requirements.txt
pywin32
python-miio